pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zkill_batch"
version = "0.1"
description = "EVE Online Killmail Processing System"
requires-python = ">=3.8"
dependencies = [
    "psycopg2-binary",
    "python-dotenv",
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]